Refactor `rlxf` to `ultralabel`
