--- src/rlxf/__init__.py
@@ -1,5 +1,0 @@
-"""
-The `rlxf` package provides functionality for generating a preference dataset using a language model.
-"""
-
-__version__ = "0.1.0"
