--- src/rlxf/prompts/__init__.py
@@ -1,3 +1,0 @@
-from rlxf.prompts.base import PromptTemplate
-
-__all__ = ["PromptTemplate"]
