+++ src/ultralabel/__init__.py
@@ -1,0 +1,5 @@
+"""
+The `ultralabel` package provides functionality for generating a preference dataset using a language model.
+"""
+
+__version__ = "0.1.0"
