+++ src/ultralabel/llm/openai_.py
@@ -11,10 +11,10 @@ from tenacity import (
     wait_random_exponential,
 )
 
-from rlxf.llm.base import LLM
+from ultralabel.llm.base import LLM
 
 if TYPE_CHECKING:
-    from rlxf.prompts.base import PromptTemplate
+    from ultralabel.prompts.base import PromptTemplate
 
 
 _OPENAI_API_RETRY_ON_EXCEPTIONS = (
