+++ src/ultralabel/prompts/__init__.py
@@ -1,0 +1,3 @@
+from ultralabel.prompts.base import PromptTemplate
+
+__all__ = ["PromptTemplate"]
