+++ src/ultralabel/prompts/templates/gpt-text-generation.jinja2
