+++ src/ultralabel/prompts/templates/gpt4-response-rating.jinja2
