+++ src/ultralabel/prompts/templates/llama2-generation.jinja2
